Hack opentofu/hashicorp namespace race condition
As we use provider aliases to fill in the HC data from opentofu, things
can get a bit weird if we are not careful.  We got unlucky with timing
causing the azurerm provider to try a version update in the same
checkout (opentofu) directory.

This is a patch while we determine the best way to actually fix this
issue.

Signed-off-by: Christian Mesh <christianmesh1@gmail.com>