Move to openapi 3.0 for spec
This commit updates the OpenAPI specification to version 3.0, which
is a more modern and widely adopted version of the OpenAPI standard.

It removes any autogeneration approach to openapi spec from golang comments
and moves towards a more manual approach because the developer experience is almost identical.

To handle swagger 2.0 users, we will generate a swagger 2.0 spec from the openapi 3.0 spec using
a script. This allows users of swagger 2.0 to continue using the API without any issues.

This also improves our redocly setup by ensuring that we're including the correct fonts
and styles, and introduces lazy loading so its not blocking the page load.

Signed-off-by: James Humphries <james@james-humphries.co.uk>